Added a new node type, "fs", which is a clone of Ram Node, except that the outbox of encrypted messages is written to disk and not to RAM. This is useful for systems with limited resources.
Due to filesystem blocking, this node has been shown to be more influenced by the polling interval than other nodes.  In tests, once a second seems stable.

There are also issues with this node due to file modified times often being identical due to cached writing by the block layer.   The current workaround is to
override the maxBytes limit from the protocol with extra messages with identical file times.  This logic will have to be improved on.

Also updated the tests to support more than two types of node and to include the new node type.