MAJOR: Added a new serialization method and made UDP use it by default.
Why not use gob anymore or protobuf? - Made an optional custom de/serialization scheme for ratnet objects
that you can OPTIONALLY use in your transports if it makes life easier.  No dependencies, and no human-
readable strings in the output.

Please note:  gob is still being used by Pickup/Dropoff at the moment, but if the internal serialization scheme
seems stable in a bit, we might consider moving that over for ease of porting nodes to other platforms in the
future.

todo: a million unit tests for specific serializations, but end-to-end unit tests all pass